Remove CredentialKey.
CredentialKey is a construct tied to one specific issuance protocol (or family
of protocols) and it's not inconceivable that other issuance protocols may want
to use different constructs. In fact, in our self-signed credentials in appholder,
CredentialKey isn't used at all.

Applications wishing to continue using a CredentialKey can simply just create it
manually using a SecureArea implementation and use Credential.applicationData to
store the alias.

This removal affects the migration code in KeystoreIdentityCredentialStore and
replaces the migrateToCredentialStore() method with getNameSpacedData() and
getCredentialKeyAlias() methods. This change actually gives the application
more flexibility insofar that it now easily maintain both the old credential
and the new credential since the old one isn't deleted at migration time.

Also change how to create keys in AndroidKeystoreSecureArea for an existing
alias.

Test: All unit tests pass.
Signed-off-by: David Zeuthen <zeuthen@google.com>