Change wallet defaults and fix race condition.
Make the wallet app default to NFC Negotiated Handover and close the
connection after sending the response. This change is to more closely
mimic wallets already in production using this code.

Also add additional checks for callback inhibition when inside the
executor to avoid transport callbacks after closing a connection. This
happened in the reader app - causing a crash - when the mDL
included status 20 in the same SessionData message which contained the
result (e.g. using auto-close).

Test: Manually tested.
Test: All unit tests pass.

Signed-off-by: David Zeuthen <zeuthen@google.com>