Adding sample for a simple age-verification reader app demonstrating (#449)
use of the MdocReaderPrompt class, which allows any app to easily act
as a reader app for the common age-verification use case. This class
abstracts much of the VerificationHelper details for a simpler API which
only requires an MdocReaderSettings object as input before it can be
shown. The MdocReaderPrompt object, once created, has a `show` method
which will show relevant UI on-screen for in-person verification. Data
from the transaction is not passed back to the app

Signed-off-by: Suzanna Jiwani <suzannaj@google.com>