wallet: Add foundations for proofing/issuance. (#465)
This includes basic types, flows, and three different evidence types as
well as an implementation of a self-signed implementation which is implemented
entirely in the application. A unit test for the happy flow is also included.

Once this code is baked it will be moved into the identity library.

This commit also includes a small extension to managedAuthenticationKeyHelper()
to perform a dry run. This is needed for the application to know how many
authentication keys will be created.

Also include a fix to get the identity-android unit tests to compile (we
really need to start running these in presubmit).

Test: Manually tested (appholder, appverifier, samples/preconsent-mdl, samples/age-verifier-mdl, wallet, secure-area-test-app)
Test: New unit tests and all unit tests pass (wallet, identity, identity-android).

Signed-off-by: David Zeuthen <zeuthen@google.com>