Add new CBOR and COSE libraries written entirely in Kotlin.
This is needed to get the libraries working with Kotlin Multiplatform.

This PR also converts all the code in identity and identity-mdoc to
use this new code instead of cbor-java and the various support
routines in the Util class. Code in identity-android and the apps will
be converted in a future PR.

Also create wrappers for all cryptographic routines needed, including
hashing, hkdf, encryption/decryption, EC keys, and certificates. This
is currently using BC but can be easily ported to other cryptographic
libraries when not running on a JVM, for example iOS.

Test: New unit tests and all unit tests pass.
Test: Manually tested (appholder, appverifier, samples/preconsent-mdl, samples/age-verifier-mdl, samples/simple-verifier, wallet, secure-area-test-app)

Signed-off-by: David Zeuthen <zeuthen@google.com>