NFC tunnel to collect evidence from a MRTD-compliant chip. DSL for issuing request/response graph. (#476)
Support new type of evidence that is collected directly by the IssuingAuthority by opening a tunnel to communicate to the chip in MRTD through NFC. If MRTD supports Chip Authentication (common case) that enables IssuingAuthority to both validate that the chip is genuine (i.e. was not cloned) and establish a secure channel of communication to the chip (that cannot be tampered by a malicious client app or other man-in-the-middle attacks).

A new pair of EvidenceRequest/EvidenceResponse types is added to carry the tunnel messages. Also an additional EvidenceResponse is defined to record the summary of the tunnel communication (in line with other evidence types, as recording tunnel communication is both verbose and impractical, as a very specialized post-processing would be required).

Also converted the grammar for request/response graph in SimpleIssuingAuthority to Kotlin DSL to improve its clarity.
Manually tested with a variety of passports

Signed-off-by: Peter Sorotokin <sorotokin@gmail.com>