Add functions and tool to create IACA and DS certificates.
This adds two new methods to `MdocUtil` for generating IACA and DS
certificates (with all the extensions required by ISO/IEC 18013-5:2021
Annex B) and also introduces a new `identityctl` command-line tool
which invokes these.

Also hook this up to the `wallet` app and add the IACA cert to
`appverifier`. The checkmarks for "Issuer DS Key Recognized" is now
green. Also leave instructions in the source for how to regenerate
these.

This includes methods to load/save keys and certificates in PEM
format. Add unit tests for this.

Also add some text to the wallet About screen and add links to our
home page and also the IACA certificate in use.

Test: New unit test and all unit tests pass.
Test: Manually tested apps.
Signed-off-by: David Zeuthen <zeuthen@google.com>