Add method MdocDataElement.renderValue() to nicely render values. (#524)
This takes the type of the value into account as per the
CredentialTypeRepository and supports enumerations used for that
particular DocType. For example, for in the AAMVA namespace instead of
rendering "1" for data element `organ_donor`, it will render
"Donor". Similarly, instead of rendering "AP" for data element
`race_ethnicity` it will render "Asian or Pacific Islander".

Use this in appverifier to display the retrieved document and test
with retrieving "Full mDL" from appholder to manually verify it's
correct. This uncovered a bug in appholder where data elements for
numbers were accidentially encoded as strings. Fix this.

This code will be used to render the values for the Credman credential
picker which has "View details" button to see the value of the data
elements which will be shared.

Test: New unit tests and all unit tests pass.
Test: Manually tested appverifier.

Signed-off-by: David Zeuthen <zeuthen@google.com>