Add support for EU PID doctype in wallet. (#542)
This includes updating our existing `EUPersonalID` type to the latest
version of the PID rulebook at

 https://github.com/eu-digital-identity-wallet/eudi-doc-architecture-and-reference-framework/blob/main/docs/annexes/annex-06-pid-rulebook.md

Slightly improve the UI in MainSreen and QrEngagementScreen. Also
rework wording in the UI so the end-user term for a credential is
"Document".

ConsentPrompt showed all requested data elements even if the selected
credential didn't have all of them. Fix this so we only show the union
of requested data and what is available.

Introduce Settings screen and consolidate with the existing "Log to
file" machinery and UI. In addition to "log to file", this exposes a
new "Developer Mode" setting. The only thing the setting does right is
that it controls whether the availability "Show keys" menu item in the
hot dog menu on the "Document" screen.

This idea behind Developer Mode is that we can use this setting in the
future to expose settings/paths useful for developers but which make
no sense to end users... for example, in the future when we have a
Wallet Server we can keep the signed-on-device issuer which is useful
for e.g. developers testing the wallet functionality at
interoperability events. This signed-on-device issuer could also use
the existing Evidence framework ask things like which authentication
or DeviceKey curve to use which is useful for developers.

Test: Manually tested.
Test: All unit tests pass.

Signed-off-by: David Zeuthen <zeuthen@google.com>