Support more error types in BiometricPrompt
In the case of no enrolled biometrics, allow users with secure lock
screen to use LSKF only to unlock auth keys. Also improves user experience
by allowing for a retry when the biometrics are not clear (ie, when the
prompt tells you to press your finger firmly against the sensor) - before,
the presentation would be cancelled after a single failure but the prompt
would remain on-screen.

Tested manually as a user with only PIN and no biometrics.

Fixes #550

Signed-off-by: Suzanna Jiwani <suzannaj@google.com>