wallet: add notification support. (#558)
This adds support for Android notifications when a document has been
proofed, when there's an update from the issuer, and when the issuer
has deleted the document.

Mark some methods/fields on Credential as internal as they shouldn't
be part of the public API.

Add a NameSpacedData.Builder construct which takes an existing
NameSpacedData which is useful if you just want to modify some fields
in an existing NameSpacedData instance.

Add `Evidence{Request, Response}NotificationPermission` evidence types
which the issuer can use to ask the user to grant notifiction
permission if the app doesn't have it already. Hook this up to
Android's POST_NOTIFICATIONS permission using the accompanist-permissions
library.

Add new monochrome app icon used in notifications.

Add new experiment icon from
https://fonts.google.com/icons?selected=Material+Symbols+Outlined:experiment:FILL@0;wght@400;GRAD@0;opsz@24&icon.query=experiment
and use it for all "Dev Mode" features to help make it clear that it's
a Dev Mode feature.

Add a new "Request Update" dev-mode feature which sends a message to
issuer to request an update. Also add the require plumbing for this
and implement it for both mDL and EU PID issuers. If called, both
issuers will update the `administrative_number` data element in their
respective namespaces.

See https://youtube.com/shorts/HBDFDWftWPU for a demo of all this new UI.

Test: Manually tested.
Test: New unit test and all unit tests pass.

Signed-off-by: David Zeuthen <zeuthen@google.com>