wallet: better handling of permissions related to proximitiy presentations.
This changes how we handle requesting proximity permissions.

Introduce a snackbar in MainScreen shown if there's at least one
credential presentation and we lack the permissions. Make the Action
button request the permission and also allow the user to dismiss
this. Record whether the snackbar was dismissed to avoid painting it
in the future via SharedPreference.

Add logic to check for permissions if "Show code" button is pressed
(for QR presentations) and if we lack permissions, show an Alert
dialog explaining that permissions need to be granted.

Add logic in `NfcEngagementHandler` to clear the preference for
whether to hide the snackbar and post a notification (best effort)
explaining to the user that they need to grant permissions. The
user clicking the notification will take them to MainScreen which
will show the snackbar even if it was previously dismissed. Also
return APPLICATION_NOT_FOUND to the remote mdoc reader so it can
take the appropriate action.

Test: Manually tested
Signed-off-by: David Zeuthen <zeuthen@google.com>