Gracefully handle when the LSKF is removed.
Introduce KeyInvalidatedException for SecureArea and mention that any
implementation may throw this expection and it means that the key is
now unusable. Also add getKeyInvalidate() method as a way to check
directly without having to use a key.

For AndroidKeystoreSecureArea, introduce ScreenLockRequiredException
and throw this if trying to create an authbound key when no screenlock
is set up.

On Document, introduce deleteInvalidatedCredentials() which uses this
facility to delete all pending and certified keys that are
invalidated. Also add hasUsableCredential() as a convenience function.

Use hot flows instead of the home-grown observer pattern on
DocumentStore, IssuingAuthority, and IssuingAuthorityRepository.

Rename CardViewModel to DocumentModel and move instantiation so it's
created by WalletApplication instead of MainActivity. This is
important because this model is used by more than just the UI. Also
rename objects exposed by DocumentModel and the composables rendering
these.

Move all credential housekeeping into DocumentModel and add support
for batch processing of the very chatty DOCUMENT_UPDATED events. This
saves a lot of processing.

Add SettingsModel.screenLockIsSetup and use this to show a snackbar
telling the user that this is the case and some things might not
work. Catch exceptions and show errors in the UI if the user is trying
to refresh a credential (w/ special-case for ScreenLockRequiredException).

In DocumentModel, add an observer for screenLockIsSetup and call
deleteInvalidatedCredentials() on all documents and make sure we
convey this to the user through the status field with "Credentials
expired or missing".

Add a worker for daily pings to the issuer (for pulling PII updates,
refreshing MSOs, etc).

Test: Manually tested
Test: All unit tests pass.
Signed-off-by: David Zeuthen <zeuthen@google.com>