Crypto: add support for HPKE encryption and decryption.
This is based on the HPKE encryption routines in CredmanUtils. It's
presently using Tink but in the future we may switch to the HPKE
implementation in BC to avoid the extra dependency (or we may
switch to using just Tink and not BC at all.)

Also port our Credman integration to use these new routines.

Test: New unit tests and all unit tests pass.
Test: Manually tested Credman presentations from appholder
  and wallet to both appverifier and https://www.identitycredential.dev/
Signed-off-by: David Zeuthen <zeuthen@google.com>