Allow device authenticaton retry. (#586)
Remove the early exit upon onAuthenticationFailed callback since it is
expected to happen occasionally e.g. when a user fingerprint isn't
recognized in the first few attempts. This allows the user to retry
biometric prompt until it reaches the built in limit, in which case
onAuthenticationError will then be invoked.

Signed-off-by: Helen Qin <helenqin@google.com>
Co-authored-by: Helen Qin <helenqin@google.com>