SecureAreaCredential and MdocCredential fixes. (#588)
The previous commit which introduced a new hierarchy of Credential
classes introduced a couple of small subtle bugs.

First problem was that MdocCredential.doctype wasn't read back
properly. This manifested itself in an error when using the "Check for
update" facility in the wallet app.

The second problem was that all SecureAreaCredential had the same
alias. This is because Credential.identifier isn't assigned until it
was added to the Document. This manifested itself in only a single
credential being visible in the wallet app.

This PR fixes these problems and also slightly refactors the code so
constructors are used instead of create(), deserialize() methods and
the fromCbor() companion object method. It also changes the flow so
the credential is added to the document from the constructor.

Test: Manually tested with wallet app.
Test: All unit tests passed.

Signed-off-by: David Zeuthen <zeuthen@google.com>