DocumentGenerator: Don't include payload in DeviceSigned COSE_Sign1.
ISO/IEC 18013-5:2021 clause 9.1.3.4 specifically said to not do
that. This bug was introduced in PR #482 when switching to the new
CBOR and COSE libraries. Fix this.

Also add a new check to Cose.coseSign1Check() for this. This check
will trigger a unit-test failure if the fix mentioned in the previous
paragraph isn't applied.

Thanks to ryanhargrove@ for finding and reporting this problem.

Test: New test and all unit tests pass.
Signed-off-by: David Zeuthen <zeuthen@google.com>