Fixes for errors when running Make Project in Android studio.
The JVM was running out of memory with 2g, so this increases that to
4g. Also, there are 3 bouncycastle libraries that all contain a
MANIFEST.MF file, and the name collision is causing an error. This
updates the build.gradle files for each of the projects that was
having an error, so it excludes this file (similar to the way it's
excluded in other modules).

Tested by running Make Project before and after, running the
wallet and appverifier apps to ensure they still work, and running
all unit tests on all modules.

Signed-off-by: Kevin Deus <kdeus@google.com>