Fix disconnected operation with Wallet Server.
Ensure the application works properly even when the Wallet Server
isn't reachable which can happen if the client isn't connected to the
Internet or the server is down. Specifically this means removing
`IssuingAuthorityRepository` and just using passing `WalletServerProvider`
around which can be used to initiate flows into the Wallet Server.

Also, since the Flow generator / library doesn't support throwing
exceptions across the network, introduce a new `NO_SUCH_DOCUMENT`
condition that can be returned in `IssuingAuthority.getState()`. This
also allows us to remove the `Document.isDeleted` property.

Add an interface to allow the wallet server to notify the client and
implement it for when the wallet server runs in the app (dev: URL).

Implement `developerModeRequestUpdate()` and verify that it works as
expected when using dev: URLs, in particular that notifications work.

Rework "Add to Wallet" screen so it calls directly into the wallet
server and make it display loading and error states as needed. The
latter is needed if trying to add a document but the wallet server
isn't reachable.

Test: Manually tested with wallet server (both dev: and across the network)
Test: All unit tests pass
Signed-off-by: David Zeuthen <zeuthen@google.com>