Move Timestamp to identity-android-legacy, replace with Instant everywhere else. (#622)
The standard kotlinx.datetime.Instant is the more common way to refer
to a specific time, rather than the custom Timestamp implementation
that this project used to use. This keeps Timestamp around for the
identity-android-legacy project but switches everything else to use
Instant.

The main non-obvious change was to move provisionSelfSignedCredential
from Utility.java to UtilTest.java, since it's one of the few
functions in identity-android-legacy that references other APIs
that were switched to Instant. The rest of the changes are fairly
mechanical.

Tested by building and running the test apps: wallet and appverifier,
and running all unit tests.

Signed-off-by: Kevin Deus <kdeus@google.com>