Add a way for the wallet server to send notifications to the wallet app.
These notifiactions only work when the appliaction is running but on
the other hand, no secrets/configuration is needed on neither the
client nor server side. Support for real push notifications using
Firebase Messaging will be added in a future commit.

Also add WalletServerCapabilities and WalletApplicationCapabilities
objects which are exchanged at authentication time, that is, the first
time a fresh wallet app instance connects to the wallet server.

Also clarify the document sync code in DocumentModel a little bit and
fix a bug there so we get notifiactions on remote document deletion.

Also enhance logging in the FlowServlet code.

Test: Manually tested.
Test: All unit tests pass.

Signed-off-by: David Zeuthen <zeuthen@google.com>