Fix network configuration problem when whitelabeling.
The CIO engine for ktor seems to have some problems, for example it
doesn't respect network_security_config.xml and it has issues when
using https URLs. Fix this by changing to the Android engine which
has none of these problems.

For the default flavor, allow any cleartext traffic and leave a note
explaining why this is allowed (so developers can easily point to
local wallet servers). For the 'customized' flavor don't allow
cleartext traffic and only allow traffic to ws.example.com. Make a
note in README.md that downstreams need to customize this.

Test: Manually tested `customized` flavor pointing to a wallet server.

Signed-off-by: David Zeuthen <zeuthen@gmail.com>