MDL Presentation Flow: Unification of Presentation Prompts
Created PresentationFlowActivity that enables any subclassing Activity
to start an MDL Presentation flow and obtain the response bytes once all
the prompts succeed. An exception is thrown otherwise.

Refactored PresentationActivity in Wallet to be a transparent Activity
that extends PresentationFlowActivity and handles showing the prompts
as well as sending the response back to the device. This allowed for
the removal of TransferHelper and cleaner code in the Activity.

Test: all unit tests ran and passed.
Test: manual MDL Presentation via NFC and QR engagement between Wallet
and AppVerifier.

Signed-off-by: dritan-x <dritan.xhabija@gmail.com>