Move server settings to web.xml. (#644)
This enables easy customization via web.xml. Also update README.md to
mention this.

This also introduces a `WalletServerSettings` class which should be
used for all settings on the Wallet Server side since it provides a
more type-safe interface. It also provides a central place to discover
what settings are available for configuration.

This will provide the foundation for Firebase Notifications which
requires both client- and server-side secrets.

Also print clientId, issuingAuthorityId, documentId when returning the
current document state. This is helpful for developers to get the values
to put into the admin web interface. Also fix a bug so this interface
works with EU PID in addition to Driving License documents.

Test: Manually tested using both dev: and wallet-server instances.

Signed-off-by: David Zeuthen <zeuthen@google.com>