mDL Presentment Flow: Unification of Presentment Prompts
Created PresentmentFlow to start a Presentment Flow for all document
requests and obtain the response bytes with all the Documents from
successful Presentment Flows (once all the prompts succeed).
An exception is thrown otherwise.

Refactored PresentationActivity in Wallet to be a transparent Activity
and removed all Compose components. Once the device listener has
the device request bytes it immediately uses PresentmentFlow to start
showing the Presentment Prompts and receive the response bytes to
send back to the device.

Test: all unit tests ran and passed.
Test: manual mDL Presentment via NFC and QR engagement between Wallet
and AppVerifier.
Test: Requesting multiple docRequests via Request mDL + micov with linkage — works for only the Mdoc type since I couldnt find a way to add a micov docType to Wallet.

Signed-off-by: dritan-x <dritan.xhabija@gmail.com>