Add selfie video support, along with ML Kit detection.
This hooks up the UI flow, video recording, and evidence propagation to
the view model and wallet server.

This also adds ML Kit, hooks it into the Selfie camera recording code, detects
all the current poses we support (and adds support for detecting smiling),
and replaces the button to proceed to the next pose.

Tested by:
- Manual tests in the wallet app
- ./gradlew check

Fixes #642
Fixes #680
Signed-off-by: Kevin Deus <kdeus@google.com>