Rename X509Certificate to X509Cert and add partial support for iOS + tests
This rename helps avoid confusion on the JVM when using
javaX509Certificate to get a java.security.cert.X509Certificate
instance for the certificate.

Also implement partial support for X509Cert on iOS - still need to
implement code for getting the EC curve from the certificate which
includes parsing ASN.1.

Test: ./gradlew check
Signed-off-by: David Zeuthen <zeuthen@google.com>