Rework iOS build system. (#653)
The use of io.github.ttypic.swiftklib causes problems when not using a
Mac for development so stop using it. Instead, include a XCode project
for the code we need on iOS. This also has the benefit that it's
easier to edit/compile/test the code in XCode plus we can add other
Swift packages as dependencies in the future. Also implement
X509Cert.ecPublicKey on iOS and move JVM-specific tests to main.

Also switch the GitHub unit test runner to using ubuntu-latest since
it's almost an order of magnitude faster than macos-latest. Once we
have wider support for iOS we might consider changing this back.

Test: ./gradlew check on both a Mac and a Linux machine.

Signed-off-by: David Zeuthen <zeuthen@google.com>