AndroidStorageEngine: Change to store data in a single file.
This is similar to the change made in GenericStorageEngine recently
and makes it simpler to write unit tests to check that new code is
backwards compatible with existing on-disk data. We want to do this so
we can guarantee backwards compat to our downstream users.

This is a breaking change, the `wallet` module will lose all
provisioned documents.

Test: ./gradlew check
Test: ./gradlew checkConnected
Test: Manually tested wallet app and secure-area-test-app.
Signed-off-by: David Zeuthen <zeuthen@google.com>