Bluetooth LE: work around bugs in BLE stack.
On some combination of devices, messages being sent to the GATT layer
are list if we call disconnect() right after sending
them. Specifically this results in the mDL reader not receiving the
entire DeviceResponse before receiving a notification that the other
device disconnected.

Work around this bug by inserting a 1 second sleep before calling
disconnect().

Test: Manually tested with two devices.

Signed-off-by: David Zeuthen <zeuthen@google.com>