BLE: Fix race condition when terminating connection immediately.
With recent changes to the code, we're terminating the BLE connection
on the mdoc side just a bit earlier which causes the mdoc reader to
sometimes receive an incomplete `SessionData` message which it fails
to decode as proper CBOR. Incomplete because it's about 500 bytes
shorter than what was sent from the mdoc. This would happen about 1 in
10 times on a setup with a Pixel 8 Pro as the mdoc and Pixel 7a as the
mdoc reader.

After digging a while it turns out this is caused by one of the GATT
writes strangely going missing. The reason for this is that
sendMessage() causes drainWritingQueue() to be called and the guarding
logic - writeIsOutstanding - didn't work. So we ended up doing two writes
to a characteristic without waiting for the other end to report that they
read for it.

Simplify all this code so we only drain the queue from sendMessage()
if it's empty to begin with.

Manually tested the fix doing around 50 transactions (NFC Negotiated
Handover + mdoc BLE central client mode), no flakes.

Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>