Add support for SD-JWT verification.
This adds support to the verifier server to request an SD-JWT PID, then check
and display the results. There are corresponding changes to the oid4vp
presentation code in the client.

This contains a number of shortcuts that should probably be refactored in
the future to make the code cleaner, including:
- Display names in the consent prompt code, rather than a new document type
  that describes all the known SD-JWT fields.
- Several "if mdoc / if sdjwt" patterns that would be cleaner if this
  functionality were moved into corresponding classes.
- Special handling for several fields that exist in the SD-JWT but are
  not in the list of disclosures.

The request and checks in the server likely have differences from the
standard. Please file bugs for any differences you find.

Tested by:
- Manual tests from an emulated device to a local verifier server.
- Manual tests from a real device to a local verifier server.
- Manual tests using our Utopia PIDs.
- Manual tests using the Funke-provided PID using SD-JWT.
- ./gradlew check

Signed-off-by: Kevin Deus <kdeus@google.com>