Adding logging for presentation activity and corresponding viewer
ActivityLogger is added and used to capture the credman presentation activity. Once an activity is made on a card, when the user view that card, they can click on the top right corner and see the corresponding activities.

Tested by: local testing for presentment activity and viewing of the event.

Signed-off-by: Hamzeh Zawawy <hamzeh@google.com>