WalletApplicationCapabilities: Convey whether app is running on an emulator.
For example, the issuer may decide to not requie a selfie check under
these conditions since this cannot be completed b/c the emulator
doesn't have a real camera.

Of course this is just a heuristic and should not be used for security
decisions (key/app attestation is the right tool for such a job).

Test: ./gradlew check
Test: ./gradlew connectedCheck

Signed-off-by: David Zeuthen <zeuthen@google.com>