MobileSecurityObjectGenerator: Don't use fractional seconds in timestamps.
ISO 18013-5 clause 9.1.2.4 specifically says to not do that and it's
desirable to be able to rely on this for some applications.

Just drop the fractional seconds and print a warning if the caller
passes in timestamps with non-zero fractional seconds.

Also add a unit test to double-check that an `Instant` serialized to a
CBOR tdate doesn't have any fractional seconds part if the `Instant`'s
`nanosecondsOfSecond` property is 0.

Test: New unit tests.
Test: All unit tests pass.
Signed-off-by: David Zeuthen <zeuthen@google.com>