Remove development team variable.
It's not good form to check this in. The only time it's needed is when
building an iOS app and for that Xcode will prompt the developer to
set the development team in the UI. Of course, if doing this then
project.pbxproj changes and care has to be taken to undo this setting
before uploading a PR.

In the future we might use add project.pbxproj.template to work around
that limitation.

Test: ./gradlew check
Test: ./gradlew connectedCheck
Test: Manually tested.
Signed-off-by: David Zeuthen <zeuthen@google.com>