Add support for Austroads Request Forwarding protocol. (#734)
This adds support for the Austroads Request Forwarding (A.R.F.)
protocol in the W3C Digital Credentials API. See "Austroads Request
Forwarding profile for the browser v4" distributed on the ISO portal
(WG10N2489).

Part of this includes a new Android Credential Matcher which supports
the "austroads-request-forwarding-v2" protocol in addition to
"preview" and "openid4vp". Another difference with this new Credman
matcher is for all three protocols, it'll match a document if just a
single of the requested data elements exist in the document, which is
consistent with the semantics defined in ISO/IEC 18013-5. Previously
the Credman matcher would match only if all the requested data
elements match the document which is fundamentally incompatible with
the mdoc data model.

Also rename ByteArray.toBase64() and String.fromBase64() extensions so
it's clear that it's Base64Url encoding and specify that padding is
omitted.

Test: Manually tested.
Test: ./gradlew check
Test: ./gradlew connectedCheck

Signed-off-by: David Zeuthen <zeuthen@google.com>