Use sample requests from DocumentTypeRepository in online verifier.
Instead of manually building UI and code, just use the sample requests
available in DocumentTypeRepository to build the UI and run the
request.

This also gives us support for the PhotoID doctype recently added. In
the future when adding new doctypes, we'll also automatically get
support in the verifier for free.

Test: Manually tested (OpenID4VP, W3C DC Preview + A.R.F., all requests)
Signed-off-by: David Zeuthen <zeuthen@google.com>