Use fewest possible bytes for CBOR length encodings.
Turns out the code for encoding a Bstr, Tstr, and other types doesn't
use the fewest possible bytes when encoding. This is allowed behavior
but the resulting CBOR is not compliant with the rules for Preferred
Serialization according to RFC 8949 4.2.1. Core Deterministic Encoding
Requirements.

This behavior triggers a bug in `DeviceResponseParser` because it's
written to assume that the encoding of `IssuerSignedItemBytes` follows
the rules in ISO/IEC 18013-5:2021 which requires this as per clause
"8.1 Encoding of data structures and data elements."

Concretely, it means that any `DeviceResponse` CBOR with bstr data
elements greater than 64 KiB will be reported as failing to validate
by `DeviceResponseParser` when in fact this is not the case. Normally
mdoc data elements are smaller than this, typically the portrait and
signature/usual-mark is only a dozen kB or so. This was only
discovered because someone at the mDL test event used a portrait image
of 90 kB.

Fix this bug and add new unit tests which ensures the fewest possible
bytes are always used for length encodings.

Also make `DeviceResponseParser` log a warning when a data element
mismatch is detected.

Test: New unit test and all unit tests pass.

Signed-off-by: David Zeuthen <zeuthen@google.com>