wallet: add new "Proximity Sharing Settings" screen. (#759)
This is a new screen only available if Developer Mode is enabled and
it allows configuration of which parts of the ISO/IEC 18013-5:2021
protocol to use. This is useful for interoperability testing with
other wallets and readers.

Also fix a bug in our implementation of _BLE mdoc peripheral server_
mode. The problem is that we forgot to start the GATT client, looks
like this bug was introduced a long time ago but was never discovered
since _BLE mdoc central client_ is the default mode for the vast
majority of wallets and readers. The fact that we now have settings to
change the default will help avoid mistakes like this in the future.

Test: Manually tested.

Signed-off-by: David Zeuthen <zeuthen@google.com>