Fix Android Studio's "Duplicate content roots detected" warning.
This resources directory is included in both the testapp main and commonMain
modules. Android Studio's response to detecting the same directory in multiple
modules is to remove it from one of them: the main module, in this case. This
change ends up with the same result: we no longer add this directory to the
main module (it's only in commonMain).

Tests still pass, so as far as I can tell this hasn't broken anything. But I
haven't been able to test iOS or any other platform than Android.

Tested by:
- ./gradlew check
- ./gradlew connectedCheck

Signed-off-by: Kevin Deus <kdeus@google.com>