identity-appsupport: Add QR code generation and scanning composables. (#769)
This adds Compose Multiplatform support for QR code generation and scanning.

A future change will start using this in the wallet app, currently
it's only used in samples/testapp.

This adds dependencies on https://github.com/alexzhirkevich/qrose and
https://github.com/kalinjul/EasyQRScan which are under the MIT and
Apache 2.0 license, respectively. These depencies are wholly hidden
behind the `ScanQrCodeDialog` and `ShowQrCodeDialog` in case we want
to swap these out in the future. For example, for the scanning part we
likely want to just use MLKit directly.

Test: New screens in samples/testapp for testing. Tested on Android and iOS.

Signed-off-by: David Zeuthen <zeuthen@google.com>