Add multiplatform BLE stack for ISO/IEC 18013-5:2021 proximity.
This adds new multiplatform API to the identity-mdoc library for
ISO/IEC 18013-5:2021 proximity presentations. It works on both Android
and iOS and both operating systems share a lot of common code.

Also add support two new pages - "ISO mdoc Proximity Sharing" and "ISO
mdoc Proximity Reading" - to the multi-platform test app. This only
includes QR engagement, NFC engagement will be added in a future
change. These two pages represent feature complete mDL and mDL reader
functionality and are designed to work with other implementations as
well as with itself.

Also introduce `rememberBluetoothPermissionState` composable function
and `BluetoothPermissionState` object which can be used to obtain the
required permissions on BLE on both Android and iOS.

Also devise a new multi-testing framework for automatically testing a
number of mDL presentations between two devices. This is available in
a new "ISO mdc Multi-Device Testing" page in the test app and the way
it works is that the 1st device offers a QR code that the 2nd device
scans. This sets up a TCP/IP control plane used for initiating tests,
including sharing `DeviceEngagement`. This helps ensure the code is
robust and well-tested both for happy paths and with all the various
termination options available in the standard.

Also modify existing L2CAP support so it follows the latest draft of
the upcoming second edition of ISO 18013-5:2021. In particular, this
means framing the `SessionData` / `SessionEstablishment` packets sent
over the L2CAP socket.

Modify `ScanQrCodeDialog` and `ShowQrCodeDialog` so they are easy to
extend.

Include portrait ans signature images in sample data in
identity-doctypes.

Test: Manually tested against wallet app and well-known readers.

Signed-off-by: David Zeuthen <zeuthen@google.com>