Improve activity viewer and log requester information
This change include changes to the activity viewer where a high level list of activities is shown. The user gets to click on each activity and can see the specific requested fields. This change required storing the requester information.

Tested by: local testing for presentment activity and viewing of the logged events.

Signed-off-by: Hamzeh Zawawy <hzawawy@gmail.com>