UI updates to Proximity Sharing and issuance completion screens.
Minor updates to the text and layout to bring them closer to the specs provided
by our UX designer.

In the case of the completion screen, the title used to mention scanning, but
many paths that reach the completion screen didn't scan anything, so I've
adjusted the text a bit.

Tested by:
- Manual testing.

Signed-off-by: Kevin Deus <kdeus@google.com>